refactor(self_improve/types): split tests to sibling tests_types.rs (#1266)
Inline test module in types.rs is moved into a sibling tests_types.rs
following the existing self_improve convention (tests_cycle, tests_history, etc.).

- src/self_improve/types.rs: 472 → 150 LOC (under 400-line rule)
- src/self_improve/tests_types.rs: new (~320 LOC)
- src/self_improve/mod.rs: register mod tests_types
- super::* → super::types::* + use crate::gym_scoring::GymSuiteScore

Pure refactor: zero logic changes, all 19 tests pass.

Closes 11/70 of #1266 (this loop iteration), 20 PRs cumulative since #1285.
